requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1

torch --extra-index-url https://download.pytorch.org/whl/cu116
torchvision --extra-index-url https://download.pytorch.org/whl/cu116

wget

matplotlib
numpy
Pillow
setuptools
pytorch_lightning
torchmetrics
scikit_learn
dvc
dvc[gdrive]

fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
wandb
